test: tidy up — remove deprecated files, share MockStateMap, parametrize mode tests
