--- tests/test_dynamics.py
@@ -1,4 +1,0 @@
-"""Deprecated: algorithm-level dynamics tests were consolidated into `test_algorithm.py`.
-
-This file intentionally contains no tests.
-"""
