+++ tests/test_smoother.py
@@ -1,4 +1,3 @@
-from typing import override
 from pathlib import Path
 from tempfile import TemporaryDirectory
 
@@ -7,22 +6,9 @@ from jax import numpy as jnp
 from omegaconf import OmegaConf, DictConfig
 import equinox as eqx
 import pytest
-from jaxfads.base import StateMap, Encoder
+from jaxfads.base import Encoder
 from jaxfads.smoother import XFADS
-
-
-class Mock(StateMap):
-    """Mock dynamics — pure deterministic transition."""
-
-    layer: eqx.Module | None
-
-    def __init__(self, conf: DictConfig, key: Array):
-        self.conf = conf
-        self.layer = None
-
-    @override
-    def eval(self, z: Array, u: Array, c: Array, *, key: Array | None = None) -> Array:
-        return z
+from conftest import MockStateMap  # noqa: F401 - class registration side-effect
 
 
 class IdentityEncoder(Encoder):
@@ -57,7 +43,7 @@ def test_constructor():
             mode="smooth",
             observation_dim=y_size,
             state_dim=z_size,
-            state_map="Mock",
+            state_map="MockStateMap",
             stepper="DiscreteStepper",
             approx="MVN",
             approx_kwargs={},
@@ -128,7 +114,7 @@ def test_top_level_dims_override_subconfig_dims():
             mode="smooth",
             observation_dim=y_size,
             state_dim=z_size,
-            state_map="Mock",
+            state_map="MockStateMap",
             stepper="DiscreteStepper",
             approx="MVN",
             approx_kwargs={},
@@ -198,147 +184,26 @@ def test_top_level_dims_override_subconfig_dims():
     assert jnp.isfinite(prior_mom).all()
 
 
-def test_low_rank_mvn_smoke_constructor_and_forward_pass():
-    """XFADS should run end-to-end with low-rank MVN encoder outputs."""
+@pytest.mark.parametrize("mode,approx_kwargs", [
+    ("smooth", {"rank": 2}),
+    ("causal", {}),
+    ("filter", {}),
+])
+def test_mode_smoke_forward_pass(mode, approx_kwargs):
+    """XFADS should run end-to-end for each standard inference mode."""
     T = 5
     y_size = 4
     z_size = 3
 
     model_conf = OmegaConf.create(
         dict(
-            mode="smooth",
+            mode=mode,
             observation_dim=y_size,
             state_dim=z_size,
-            state_map="Mock",
+            state_map="MockStateMap",
             stepper="DiscreteStepper",
             approx="MVN",
-            approx_kwargs={"rank": 2},
-            mc_size=2,
-            seed=0,
-            n_steps=T,
-            fb_penalty=0,
-            noise_penalty=0,
-            dropout=0.0,
-            dyn_conf=OmegaConf.create(
-                dict(
-                    input_dim=0,
-                    context_dim=0,
-                    state_noise=1.0,
-                    system_type="discrete",
-                )
-            ),
-            enc_conf=OmegaConf.create(
-                dict(
-                    width=8,
-                    depth=1,
-                    dropout=0.0,
-                )
-            ),
-            obs_conf=OmegaConf.create(
-                dict(
-                    model="GLM",
-                    emission_noise=1.0,
-                    norm_readout=False,
-                    dropout=0.0,
-                    likelihood="Poisson",
-                )
-            ),
-        )
-    )
-
-    key = jr.key(0)
-    model = XFADS(model_conf, key)
-
-    times = jnp.broadcast_to(jnp.arange(T), (1, T))
-    y = jr.poisson(jr.key(1), jnp.ones((1, T, y_size)))
-    u = jnp.zeros((1, T, 0))
-    c = jnp.zeros((1, T, 0))
-
-    model = model.initialize(times, y, u, c)
-    _, post_mom, prior_mom = model(times, y, u, c, key=jr.key(2))
-
-    assert jnp.isfinite(post_mom).all()
-    assert jnp.isfinite(prior_mom).all()
-
-
-def test_causal_mode_smoke_constructor_and_forward_pass():
-    """XFADS should run end-to-end with mode='causal'."""
-    T = 5
-    y_size = 4
-    z_size = 3
-
-    model_conf = OmegaConf.create(
-        dict(
-            mode="causal",
-            observation_dim=y_size,
-            state_dim=z_size,
-            state_map="Mock",
-            stepper="DiscreteStepper",
-            approx="MVN",
-            approx_kwargs={},
-            mc_size=2,
-            seed=0,
-            n_steps=T,
-            fb_penalty=0,
-            noise_penalty=0,
-            dropout=0.0,
-            dyn_conf=OmegaConf.create(
-                dict(
-                    input_dim=0,
-                    context_dim=0,
-                    state_noise=1.0,
-                    system_type="discrete",
-                )
-            ),
-            enc_conf=OmegaConf.create(
-                dict(
-                    width=8,
-                    depth=1,
-                    dropout=0.0,
-                )
-            ),
-            obs_conf=OmegaConf.create(
-                dict(
-                    model="GLM",
-                    emission_noise=1.0,
-                    norm_readout=False,
-                    dropout=0.0,
-                    likelihood="Poisson",
-                )
-            ),
-        )
-    )
-
-    key = jr.key(0)
-    model = XFADS(model_conf, key)
-
-    times = jnp.broadcast_to(jnp.arange(T), (1, T))
-    y = jr.poisson(jr.key(1), jnp.ones((1, T, y_size)))
-    u = jnp.zeros((1, T, 0))
-    c = jnp.zeros((1, T, 0))
-
-    model = model.initialize(times, y, u, c)
-    _, post_mom, prior_mom = model(times, y, u, c, key=jr.key(2))
-
-    assert jnp.isfinite(post_mom).all()
-    assert jnp.isfinite(prior_mom).all()
-
-
-def test_filter_mode_smoke_constructor_and_forward_pass():
-    """XFADS should run end-to-end with mode='filter'."""
-    T = 5
-    y_size = 4
-    z_size = 3
-
-    model_conf = OmegaConf.create(
-        dict(
-            mode="filter",
-            observation_dim=y_size,
-            state_dim=z_size,
-            state_map="Mock",
-            stepper="DiscreteStepper",
-            approx="MVN",
-            approx_kwargs={},
+            approx_kwargs=approx_kwargs,
             mc_size=2,
             seed=0,
             n_steps=T,
@@ -397,7 +262,7 @@ def test_invalid_mode_error_lists_filter_smooth_causal():
             mode="unknown",
             observation_dim=y_size,
             state_dim=z_size,
-            state_map="Mock",
+            state_map="MockStateMap",
             stepper="DiscreteStepper",
             approx="MVN",
             approx_kwargs={},
@@ -450,7 +315,7 @@ def test_filter_mode_skips_beta_encoder():
             mode="filter",
             observation_dim=y_size,
             state_dim=z_size,
-            state_map="Mock",
+            state_map="MockStateMap",
             stepper="DiscreteStepper",
             approx="MVN",
             approx_kwargs={},
@@ -512,7 +377,7 @@ def test_xfads_nofilt_mode():
             mode="nofilt",
             observation_dim=y_size,
             state_dim=z_size,
-            state_map="Mock",
+            state_map="MockStateMap",
             stepper="DiscreteStepper",
             approx="MVN",
             approx_kwargs={},
