--- tests/test_vi.py
@@ -1,4 +1,0 @@
-"""Deprecated: ELBO tests were consolidated into `test_algorithm.py`.
-
-This file intentionally contains no tests.
-"""
